feat : Add Facebook API and configure setting for using it        Add facebook login/logout function

+++ OnTheMap/StudentMapViewController.swift
@@ -8,9 +8,13 @@
 
 import UIKit
 import MapKit
+import FBSDKCoreKit
+import FBSDKLoginKit
 
 
 class StudentMapViewController : UIViewController, MKMapViewDelegate {
+    
+    // MARK: - Properties
     
     @IBOutlet weak var mapView: MKMapView!
     @IBOutlet weak var navigationBar: UINavigationBar!
@@ -73,18 +77,26 @@ class StudentMapViewController : UIViewController, MKMapViewDelegate {
     
     @IBAction func logoutButtonClicked(sender: AnyObject) {
         
-        UdacityClient.sharedInstance().deleteSession() {  success , errorString in
+        if ( FBSDKAccessToken.currentAccessToken() != nil ) {
+            FBSDKLoginManager().logOut()
+            FBSDKAccessToken.setCurrentAccessToken(nil)
+            dispatch_async(dispatch_get_main_queue(), {
+                self.dismissViewControllerAnimated(true, completion: nil)
+            })
+        } else {
+            UdacityClient.sharedInstance().deleteSession() {  success , errorString in
             
-            if success {
-                // If implementing 'dismissViewControllerAnimated' function without dispatch_async block, an error occurs with the following message
-                // 'This application is modifying the autolayout engine from a background thread, which can lead to engine corruption and weird crashes.  This will cause an exception in a future release.'
-                dispatch_async(dispatch_get_main_queue(), {
-                    self.dismissViewControllerAnimated(true, completion: nil)
-                })
-            } else {
-                print(errorString)
+                if success {
+                    // If implementing 'dismissViewControllerAnimated' function without dispatch_async block, an error occurs with the following message
+                    // 'This application is modifying the autolayout engine from a background thread, which can lead to engine corruption and weird crashes.  This will cause an exception in a future release.'
+                    dispatch_async(dispatch_get_main_queue(), {
+                        self.dismissViewControllerAnimated(true, completion: nil)
+                    })
+                } else {
+                    print(errorString)
+                }
+            
             }
-            
         }
     }
